+++ packages/ui/src/components/organisms/Form/DynamicForm/DynamicForm.unit.test.tsx
@@ -92,6 +92,7 @@ describe('DynamicFormV2', () => {
     onSubmit: vi.fn(),
     onEvent: vi.fn(),
     metadata: {},
+    disabled: false,
   } as unknown as IDynamicFormProps<any>;
 
   it('should render without crashing', () => {
@@ -125,6 +126,7 @@ describe('DynamicFormV2', () => {
     render(<DynamicFormV2 {...mockProps} />);
     expect(useValues).toHaveBeenCalledWith({
       values: mockProps.values,
+      schema: mockProps.elements,
       onChange: mockProps.onChange,
       onFieldChange: mockProps.onFieldChange,
     });
@@ -197,6 +199,7 @@ describe('DynamicFormV2', () => {
       },
       metadata: mockProps.metadata,
       validationParams: mockProps.validationParams,
+      disabled: mockProps.disabled,
     });
   });