+++ services/workflows-service/src/workflow/workflow-runtime-data.repository.ts
@@ -95,6 +95,12 @@ export class WorkflowRuntimeDataRepository {
   async findByIdWithRelations(id: string, projectIds: TProjectIds) {
     assertIsValidProjectIds(projectIds);
 
+    const test = () => {
+      return `
+
+      `;
+    };
+
     const [parentWorkflow, ...childWorkflows] = (await this.prismaService.$queryRaw`
             with workflow as (
         select wrd.id,