+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/ui
 
+## 0.7.132
+
+### Patch Changes
+
+- Bump
+- Added defaultValue to field definitions
+
 ## 0.7.130
 
 ### Patch Changes